HEC-31: Add computed attributes to CLI inspector and MCP tooling
Fill two remaining gaps in the computed attributes feature: hecks inspect
now displays a Computed Attributes section, and the MCP server exposes
an add_computed tool for AI agents. Extracted SecondaryFormatters and
AggregateLifecycleTools modules to keep files under the 200-line limit.
