+++ hecks_ai/lib/hecks_ai/aggregate_tools.rb
@@ -15,9 +15,10 @@ module Hecks
     #   - +add_entity+        -- add a sub-entity with identity to an aggregate
     #   - +add_validation+    -- add a validation rule to an aggregate field
     #   - +add_policy+        -- add a reactive policy (event -> trigger)
-    #   - +add_lifecycle+     -- add a state machine to an aggregate
-    #   - +add_transition+    -- add a lifecycle transition
-    #   - +remove_aggregate+  -- remove an aggregate from the domain
+    #   - +add_lifecycle+     -- see AggregateLifecycleTools
+    #   - +add_transition+    -- see AggregateLifecycleTools
+    #   - +add_computed+      -- see AggregateLifecycleTools
+    #   - +remove_aggregate+  -- see AggregateLifecycleTools
     #
     module AggregateTools
       # Registers all aggregate structure tools on the given MCP server.
@@ -188,54 +189,7 @@ module Hecks
           end
         end
 
-        server.define_tool(
-          name: "add_lifecycle",
-          description: "Add a state machine (e.g. status: draft -> published -> archived)",
-          input_schema: {
-            type: "object",
-            properties: {
-              aggregate: { type: "string" },
-              field: { type: "string", description: "Attribute that holds the state (e.g. status)" },
-              default: { type: "string", description: "Initial state (e.g. draft)" }
-            },
-            required: ["aggregate", "field", "default"]
-          }
-        ) do |args|
-          ctx.ensure_session!
-          ctx.capture_output do
-            handle = ctx.workshop.aggregate(args["aggregate"])
-            handle.lifecycle(args["field"].to_sym, default: args["default"])
-          end
-        end
-
-        server.define_tool(
-          name: "add_transition",
-          description: "Add a lifecycle transition (e.g. PublishPost -> published)",
-          input_schema: {
-            type: "object",
-            properties: {
-              aggregate: { type: "string" },
-              command: { type: "string", description: "Command name (e.g. PublishPost)" },
-              target: { type: "string", description: "Target state (e.g. published)" }
-            },
-            required: ["aggregate", "command", "target"]
-          }
-        ) do |args|
-          ctx.ensure_session!
-          ctx.capture_output do
-            handle = ctx.workshop.aggregate(args["aggregate"])
-            handle.transition(args["command"] => args["target"])
-          end
-        end
-
-        server.define_tool(
-          name: "remove_aggregate",
-          description: "Remove a thing from the domain",
-          input_schema: { type: "object", properties: { name: { type: "string" } }, required: ["name"] }
-        ) do |args|
-          ctx.ensure_session!
-          ctx.capture_output { ctx.workshop.remove(args["name"]) }
-        end
+        AggregateLifecycleTools.register(server, ctx)
       end
     end
   end
