+++ hecks_ai/lib/hecks_ai/mcp_server.rb
@@ -2,6 +2,7 @@ require "mcp"
 require_relative "type_resolver"
 require_relative "session_tools"
 require_relative "aggregate_tools"
+require_relative "aggregate_lifecycle_tools"
 require_relative "inspect_tools"
 require_relative "build_tools"
 require_relative "play_tools"
