+++ hecksties/lib/hecks_cli/domain_inspector/aggregate_formatter.rb
@@ -3,15 +3,19 @@
 # Formats a single aggregate from the domain IR into readable terminal output.
 # Covers attributes, value objects, entities, lifecycle, commands, events,
 # queries, validations, invariants, policies, scopes, specifications,
-# subscribers, and references.
+# subscribers, computed attributes, and references.
 #
 #   formatter = AggregateFormatter.new(aggregate)
 #   formatter.format  # => Array<String>
 #
+require_relative "secondary_formatters"
+
 module Hecks
   class CLI
     class DomainInspector
       class AggregateFormatter
+        include SecondaryFormatters
+
         # @param agg [Hecks::DomainModel::Structure::Aggregate]
         def initialize(agg)
           @agg = agg
@@ -24,6 +28,7 @@ module Hecks
           lines << "=" * (11 + @agg.name.length)
           lines << ""
           lines.concat(format_attributes)
+          lines.concat(format_computed_attributes)
           lines.concat(format_value_objects)
           lines.concat(format_entities)
           lines.concat(format_lifecycle)
@@ -166,44 +171,6 @@ module Hecks
             body = Hecks::Utils.block_source(pol.block)
             "    #{pol.name}: guard#{async_note} — #{body}"
           end
-        end
-
-        def format_scopes
-          return [] if @agg.scopes.empty?
-          lines = ["  Scopes:"]
-          @agg.scopes.each { |s| lines << "    #{s.name}" }
-          lines << ""
-        end
-
-        def format_specifications
-          return [] if @agg.specifications.empty?
-          lines = ["  Specifications:"]
-          @agg.specifications.each do |s|
-            body = Hecks::Utils.block_source(s.block)
-            lines << "    #{s.name}: #{body}"
-          end
-          lines << ""
-        end
-
-        def format_subscribers
-          return [] if @agg.subscribers.empty?
-          lines = ["  Subscribers:"]
-          @agg.subscribers.each do |s|
-            async_note = s.async ? " [async]" : ""
-            body = Hecks::Utils.block_source(s.block)
-            lines << "    #{s.name}: on #{s.event_name}#{async_note} — #{body}"
-          end
-          lines << ""
-        end
-
-        def format_references
-          return [] if @agg.references.empty?
-          lines = ["  References:"]
-          @agg.references.each do |ref|
-            kind = ref.respond_to?(:kind) && ref.kind ? " (#{ref.kind})" : ""
-            lines << "    -> #{ref.type}#{kind}"
-          end
-          lines << ""
         end
       end
     end
